// mobile/src/domain/database.types.ts
export interface Database {
  public: {
// ...
    Views: {
      GameResult: {
        Row: {
          categoryId: number | null;
          clubId: number | null;
          count: number | null;
          date: string | null;
          duration: number | null;
          external: boolean | null;
          gameEnded: boolean | null;
          id: number | null;
          name: string | null;
          teamGameId: number | null;
          teamName: string | null;
          type: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "Game_categoryId_fkey";
            columns: ["categoryId"];
            referencedRelation: "Category";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "Game_clubId_fkey";
            columns: ["clubId"];
            referencedRelation: "Club";
            referencedColumns: ["id"];
          },
        ];
      };
      PlayerScore: {
        Row: {
          count: number | null;
          memberId: number | null;
          type: string | null;
        };
        Relationships: [];
      };
      TeamScore: {
        Row: {
          count: number | null;
          gameId: number | null;
          teamGameId: number | null;
          type: string | null;
        };
        Relationships: [
          {
            foreignKeyName: "TeamGame_gameId_fkey";
            columns: ["gameId"];
            referencedRelation: "Game";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "TeamGame_gameId_fkey";
            columns: ["gameId"];
            referencedRelation: "GameResult";
            referencedColumns: ["id"];
          },
        ];
      };
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
}